+++ OpCode.h
@@ -2,7 +2,7 @@
 
 class OpCode{
     public:
-        OpCode(unsigned short high_byte, unsigned short low_byte);
+        OpCode(uint8_t high_byte, uint8_t low_byte);
 
         uint8_t High();
         uint8_t Low();
